[update] refs #53 パス名を定義。ルーティング変更
 UI/UX改善のため
